fixed filter switch

+++ web/search.py
@@ -40,7 +40,8 @@ def change_search_filter(search_type):
 
     # dropping old index and creating new one for each search
     # (mongodb allows only 1 text index per collection)
-    collection.drop_index("text_index")
+    if len(list(collection.list_indexes())) != 1:
+        collection.drop_index("text_index")
 
     # all possible options for index and weights
     text_indexes = dict(
